refactor: code package refactoring
