+++ internal/components/table.go
@@ -1,0 +1,7 @@
+package components
+
+type Table interface {
+	Columns() []string
+	Rows() []string
+	OnTableResize(int)
+}
