+++ internal/ui/handler_search.go
@@ -1,0 +1,12 @@
+package ui
+
+import (
+	"log"
+	"time"
+)
+
+func HandleSearch(u *Ui) {
+	log.Println("TODO Search Instance")
+	time.Sleep(time.Second * 2) // Give user some seconds to read the error
+	u.Render()
+}
