+++ internal/ui/interface.go
@@ -1,0 +1,11 @@
+package ui
+
+import (
+	"github.com/gdamore/tcell/v2"
+)
+
+type UiInterface interface {
+	Render()
+	Run() error
+	GetScreen() tcell.Screen
+}
